Time: 278 ms (12.90%), Space: 46.1 MB (38.69%) - LeetHub

+++ 985-sum-of-even-numbers-after-queries/985-sum-of-even-numbers-after-queries.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     vector<int> sumEvenAfterQueries(vector<int>& nums, vector<vector<int>>& queries) {
         vector<int> ans;
-        long long sum =0;
+        long sum =0;
         
          int n = nums.size();
